+++ src/applications/vaos/reducers/expressCare.js
@@ -60,29 +60,11 @@ export default function expressCareReducer(state = initialState, action) {
         maxEnd = timesReverseSorted?.[0];
       }
 
-      const expressCare = {
-        windowsStatus: FETCH_STATUS.succeeded,
-        allowRequests:
-          times.length && nowUtc.isBetween(minStart?.utcStart, maxEnd?.utcEnd),
-        minStart,
-        maxEnd,
-        localWindowString:
-          minStart && maxEnd
-            ? `${moment
-                .parseZone(minStart.start)
-                .format('h:mm a')} to ${moment
-                .parseZone(maxEnd.end)
-                .format('h:mm a')} ${minStart.timeZone}`
-            : null,
-      };
-
       return {
         ...state,
         windowsStatus: FETCH_STATUS.succeeded,
         allowRequests:
-          times.length &&
-          nowUtc.isAfter(moment.utc(minStart?.utcStart)) &&
-          nowUtc.isBefore(moment.utc(maxEnd?.utcEnd)),
+          times.length && nowUtc.isBetween(minStart?.utcStart, maxEnd?.utcEnd),
         minStart,
         maxEnd,
         localWindowString:
