+++ src/applications/vaos/tests/reducers/appointments.unit.spec.js
@@ -345,55 +345,4 @@ describe('VAOS reducer: appointments', () => {
     expect(newState.futureStatus).to.equal(FETCH_STATUS.notStarted);
     expect(newState.future).to.be.null;
   });
-
-  describe('express care window', () => {
-    it('should set windowsStatus to loading', () => {
-      const action = {
-        type: FETCH_EXPRESS_CARE_WINDOWS,
-      };
-
-      const newState = appointmentsReducer(initialState, action);
-      expect(newState.expressCare.windowsStatus).to.equal(FETCH_STATUS.loading);
-    });
-
-    it('should fetch and format express care window and update windowsStatus', () => {
-      const window = {
-        start: '00:00',
-        end: '23:59',
-        timezone: 'MDT',
-        offsetUtc: '-06:00',
-      };
-      const action = {
-        type: FETCH_EXPRESS_CARE_WINDOWS_SUCCEEDED,
-        facilityData: [
-          [
-            {
-              expressTimes: window,
-            },
-          ],
-        ],
-        nowUtc: moment.utc(),
-      };
-
-      const newState = appointmentsReducer(initialState, action);
-      const { expressCare } = newState;
-      expect(expressCare.windowsStatus).to.equal(FETCH_STATUS.succeeded);
-      expect('allowRequests' in expressCare).to.equal(true);
-      const today = moment.utc();
-      const startString = `${today}T${window.start}${window.offsetUtc}`;
-      const endString = `${today}T${window.start}${window.offsetUtc}`;
-      expect(expressCare.localWindowString).to.equal(
-        '12:00 a.m. to 11:59 p.m. MDT',
-      );
-    });
-
-    it('should set windowsStatus to failed', () => {
-      const action = {
-        type: FETCH_EXPRESS_CARE_WINDOWS_FAILED,
-      };
-
-      const newState = appointmentsReducer(initialState, action);
-      expect(newState.expressCare.windowsStatus).to.equal(FETCH_STATUS.failed);
-    });
-  });
 });
